pyproject.toml
[project]
name = "gptbundlev2"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "fastapi[standard]>=0.124.0",
    "litellm>=1.80.10",
    "passlib[argon2]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pynamodb>=6.1.0",
    "sqlmodel>=0.0.27",
    "websockets>=15.0.1",
    "typer>=0.15.1",
    "rich>=13.9.4",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.8.4",
]

[tool.ruff]
line-length = 88
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4", "PT"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["gptbundle*"]

[project.scripts]
admin-cli = "gptbundle.cli:app"
